refactor(rebrand): consolidated into 'main' package
added README.md
