--- .github/workflows/publish.yaml
@@ -1,27 +1,0 @@
-name: Publish Package to npmjs
-on:
-  push:
-    branches: [master]
-jobs:
-  publish:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - uses: pnpm/action-setup@v2
-        with:
-          version: 7
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v3
-        with:
-          node-version: ${{ matrix.node-version }}
-          cache: 'pnpm'
-
-      - name: Install dependencies
-        run: pnpm install
-
-      - run: pnpm -r check-format
-      - run: pnpm -r test
-      - run: pnpm -r build
-      - run: pnpm -r publish
-        env:
-          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
