--- packages/types/CHANGELOG.md
@@ -1,25 +1,0 @@
-# @pmaltese/prisma-schema-dsl-types
-
-## 1.2.1
-
-### Patch Changes
-
-- e8acb3c: The tests were split into multiple files and the previous tests abstraction was removed (the 'cases' array).
-
-## 1.2.0
-
-### Minor Changes
-
-- Datasource: added 'relationMode' field
-
-## 1.1.0
-
-### Minor Changes
-
-- Implemented: generator preview features, model indexes and fullText indexes
-
-## 1.0.10
-
-### Patch Changes
-
-- 80cdf80: Refactor: monorepo setup
