--- packages/types/README.md
@@ -1,2 +1,0 @@
-# prisma-schema-dsl-types
-types library for [prisma-schema-dsl](https://github.com/amplication/prisma-schema-dsl)
