--- packages/types/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@pmaltese/prisma-schema-dsl-types",
-  "version": "1.2.1",
-  "description": "types library for prisma-schema-dsl",
-  "main": "dist/index.js",
-  "scripts": {
-    "build": "tsc",
-    "format": "prettier --write src",
-    "check-format": "prettier --check src",
-    "prepare": "pnpm run build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/amplication/prisma-schema-dsl-types.git"
-  },
-  "keywords": [
-    "prisma",
-    "schema",
-    "dsl",
-    "javascript",
-    "ast",
-    "generator"
-  ],
-  "author": {
-    "name": "Amplication",
-    "email": "contact@amplication.com",
-    "url": "https://amplication.com"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "typescript": "^4.9.3"
-  },
-  "devDependencies": {
-    "prettier": "^2.8.0"
-  }
-}
