--- packages/types/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "strict": true,
-    "declaration": true,
-    "removeComments": true,
-    "target": "es3",
-    "sourceMap": true,
-    "outDir": "./dist",
-    "baseUrl": "./src",
-    "incremental": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist"]
-}
